Changing name plates based on player input
